backend/requirements.txt
Flask
pdfplumber
pytesseract
pandas
spacy
Pillow
pillow-heif
pyheif
